create-account/requirements.txt
psycopg2-binary
pyotp
qrcode[pil]
cryptography
Flask